Add name to user. Move classes from data to model package.
