--- src/main/kotlin/com/communicator/the/data/User.kt
@@ -1,5 +1,0 @@
-package com.communicator.the.data
-
-class User(
-        val id: String
-)
