+++ src/main/kotlin/com/communicator/the/model/User.kt
@@ -1,0 +1,6 @@
+package com.communicator.the.model
+
+class User(
+        val id: String?=null,
+        val name: String? = null
+)
