+++ src/main/kotlin/com/communicator/the/model/UserWebSocketSession.kt
@@ -1,4 +1,4 @@
-package com.communicator.the.data
+package com.communicator.the.model
 
 class UserWebSocketSession(
         val sessionId: String,
